Move all icloud export/import features to new Download modal
