+++ src/actions/actionTypes.js
@@ -6,6 +6,7 @@ export const OPEN_MODAL = 'routing/open_modal';
 export const CLOSE_MODAL = 'routing/close_modal';
 export const TOGGLE_SETTINGS = 'routing/toggle_settings';
 export const TOGGLE_ABOUT = 'routing/toggle_about';
+export const TOGGLE_DOWNLOAD = 'routing/toggle_download';
 export const HIDE_LOADING = 'routing/hide_loading';
 export const TOGGLE_TIMER = 'routing/toggle_timer';
