+++ src/helpers/icons.js
@@ -14,6 +14,7 @@ export const addShowIcon = (<MaterialIcon name="group-add" size={25} color="#FFF
 export const addIcon = (<MaterialIcon name="add" size={25} color="#000" style={statusBarStyles.navLinkIcon} />);
 export const settingsIcon = (<MaterialIcon name="settings" size={25} color="#000" style={statusBarStyles.navLinkIcon} />);
 export const aboutIcon = (<MaterialIcon name="help-outline" size={25} color="#000" style={statusBarStyles.navLinkIcon} />);
+export const downloadIcon = (<MaterialIcon name="file-download" size={25} color="#000" style={statusBarStyles.navLinkIcon} />);
 export const menuIcon = (<MaterialIcon name="menu" size={32} color="#000" style={statusBarStyles.navLinkIcon} />);
 export const closeIcon = (<MaterialIcon name="close" size={32} color="#000" style={statusBarStyles.navLinkIcon} />);
 export const bulletIcon = (<MaterialIcon name="fiber-manual-record" size={10} color="#000" style={statusBarStyles.navLinkIcon} />);
@@ -47,6 +48,7 @@ export default {
   addIcon,
   settingsIcon,
   aboutIcon,
+  downloadIcon,
   menuIcon,
   closeIcon,
   bulletIcon,
